+++ packages/contracts/hardhat.config.ts
@@ -33,8 +33,6 @@ extendEnvironment(hre => {
   const aragonPluginRepos: AragonPluginRepos = {
     AddresslistVotingRepoProxy: '',
     TokenVotingRepoProxy: '',
-    AdminRepoProxy: '',
-    MultisigRepoProxy: '',
   };
   const testingFork: TestingFork = {
     network: '',
@@ -43,7 +41,7 @@ extendEnvironment(hre => {
   };
   hre.aragonPluginRepos = aragonPluginRepos;
   hre.aragonToVerifyContracts = [];
-  hre.managementDAOMultisigPluginAddress = '';
+  hre.managementDAOMultisigPluginAddress = ''; // TODO
   hre.managementDAOActions = [];
   hre.testingFork = testingFork;
 });