+++ packages/contracts/test/deploy/managing-dao.ts
@@ -55,8 +55,6 @@ describe('Management DAO', function () {
   let pluginsRepos: {
     tokenVoting: PluginRepo;
     addresslistVoting: PluginRepo;
-    admin: PluginRepo;
-    multisig: PluginRepo;
   };
 
   before(async () => {
@@ -141,14 +139,6 @@ describe('Management DAO', function () {
       ),
       addresslistVoting: PluginRepo__factory.connect(
         hre.aragonPluginRepos.AddresslistVotingRepoProxy,
-        deployer
-      ),
-      admin: PluginRepo__factory.connect(
-        hre.aragonPluginRepos.AdminRepoProxy,
-        deployer
-      ),
-      multisig: PluginRepo__factory.connect(
-        hre.aragonPluginRepos.MultisigRepoProxy,
         deployer
       ),
     };
@@ -296,53 +286,6 @@ describe('Management DAO', function () {
         expect(
           await pluginsRepos.addresslistVoting.isGranted(
             pluginsRepos.addresslistVoting.address,
-            managementDao.address,
-            PLUGIN_REPO_PERMISSIONS.MAINTAINER_PERMISSION_ID,
-            []
-          )
-        ).to.be.true;
-      });
-    });
-
-    context('Admin Repo', async function () {
-      it('has permission to upgrade the repo', async function () {
-        expect(
-          await pluginsRepos.admin.isGranted(
-            pluginsRepos.admin.address,
-            managementDao.address,
-            PLUGIN_REPO_PERMISSIONS.UPGRADE_REPO_PERMISSION_ID,
-            []
-          )
-        ).to.be.true;
-      });
-      it('has permission to maintain the repo', async function () {
-        expect(
-          await pluginsRepos.admin.isGranted(
-            pluginsRepos.admin.address,
-            managementDao.address,
-            PLUGIN_REPO_PERMISSIONS.MAINTAINER_PERMISSION_ID,
-            []
-          )
-        ).to.be.true;
-      });
-    });
-
-    context('Multisig Repo', async function () {
-      it('has permission to upgrade the repo', async function () {
-        expect(
-          await pluginsRepos.multisig.isGranted(
-            pluginsRepos.multisig.address,
-            managementDao.address,
-            PLUGIN_REPO_PERMISSIONS.UPGRADE_REPO_PERMISSION_ID,
-            []
-          )
-        ).to.be.true;
-      });
-
-      it('has permission to maintain the repo', async function () {
-        expect(
-          await pluginsRepos.multisig.isGranted(
-            pluginsRepos.multisig.address,
             managementDao.address,
             PLUGIN_REPO_PERMISSIONS.MAINTAINER_PERMISSION_ID,
             []