+++ packages/contracts/CHANGELOG.md
@@ -26,6 +26,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
+- Moved `Multisig`, `MultisigSetup` and related metadata to the [aragon/multisig-plugin](https://github.com/aragon/multisig-plugin) repo.
+- Moved `Admin`, `AdminSetup` and related metadata to the [aragon/admin-plugin](https://github.com/aragon/admin-plugin) repo.
 - Moved
 
   - interfaces `IDAO`, `IPermissionCondition`, `IPlugin`, `IMembership`, `IProposal`, `IPluginSetup`, `IProtocolVersion`,